Rename to lower case.
